// types/enums.ts
export enum MindMapItemColors {
  RED,
  GREEN,
  PURPLE,
  BLUE,
  CYAN,
  EMERALD,
  INDIGO,
  LIME,
  ORANGE,
  PINK,
  YELLOW,
  DEFAULT,
  FUCHSIA,
}

export enum EdgeColor {
  RED = "RED",
  GREEN = "GREEN",
  PURPLE = "PURPLE",
  BLUE = "BLUE",
  CYAN = "CYAN",
  EMERALD = "EMERALFD",
  INDIGO = "INDIGO",
  LIME = "LIME",
  ORANGE = "ORANGE",
  PINK = "PINK",
  YELLOW = "YELLOW",
  DEFAULT = "DEFAULT",
  FUCHSIA = "FUCHSIA",
}